feat: vault swap => vault migrator + basic testing
