+++ scripts/deploy_vault_swap.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from brownie import VaultSwap, accounts, config, network, project, web3
+from brownie import VaultMigrator, accounts, config, network, project, web3
 from eth_utils import is_checksum_address
 
 
@@ -9,7 +9,7 @@ def main():
     dev = accounts.load("dev")
     print(f"You are using: 'dev' [{dev.address}]")
 
-    if input("Deploy VaultSwap? y/[N]: ").lower() != "y":
+    if input("Deploy VaultMigrator? y/[N]: ").lower() != "y":
         return
 
-    vaultSwap = VaultSwap.deploy({"from": dev})
+    VaultMigrator.deploy({"from": dev})
